Add support for auto-package discovery. Refactor service provider and facade.
