+++ composer.json
@@ -1,32 +1,50 @@
 {
-    "name": "yajra/zillow",
-    "type": "application",
-    "description": "Laravel Wrapper for Zillow API",
-    "keywords": ["laravel", "zillow", "php", "api"],
-    "homepage": "https://github.com/yajra/zillow",
-    "license": "MIT",
-    "authors": [
-        {
-            "name": "Arjay Angeles",
-            "email": "aqangeles@gmail.com"
-        }
-    ],
-    "support": {
-        "issues": "https://github.com/yajra/zillow/issues"
+  "name": "yajra/zillow",
+  "type": "application",
+  "description": "Laravel Wrapper for Zillow API",
+  "keywords": [
+    "laravel",
+    "zillow",
+    "php",
+    "api"
+  ],
+  "homepage": "https://github.com/yajra/zillow",
+  "license": "MIT",
+  "authors": [
+    {
+      "name": "Arjay Angeles",
+      "email": "aqangeles@gmail.com"
+    }
+  ],
+  "support": {
+    "issues": "https://github.com/yajra/zillow/issues"
+  },
+  "require": {
+    "php": ">=5.4.0",
+    "illuminate/support": "5.*",
+    "guzzlehttp/guzzle": "4.*|5.*|6.*",
+    "fabpot/goutte": "~2.0"
+  },
+  "require-dev": {
+    "phpunit/phpunit": "4.*"
+  },
+  "extra": {
+    "branch-alias": {
+      "dev-master": "3.0-dev"
     },
-    "require": {
-        "php": ">=5.4.0",
-        "illuminate/support": "5.*",
-        "guzzlehttp/guzzle": "4.*|5.*|6.*",
-        "fabpot/goutte": "~2.0"
-    },
-    "require-dev": {
-        "phpunit/phpunit": "4.*"
-    },
-    "autoload": {
-        "psr-0": {
-            "Yajra\\Zillow": "src/"
-        }
-    },
-    "minimum-stability": "stable"
+    "laravel": {
+      "providers": [
+        "Yajra\\Zillow\\ZillowServiceProvider"
+      ],
+      "aliases": {
+        "Zillow": "Yajra\\Zillow\\Facades\\Zillow"
+      }
+    }
+  },
+  "autoload": {
+    "psr-0": {
+      "Yajra\\Zillow": "src/"
+    }
+  },
+  "minimum-stability": "stable"
 }
