+++ src/yajra/Zillow/Facades/Zillow.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Yajra\Zillow;
+namespace Yajra\Zillow\Facades;
 
-use Illuminate\Support\Facades\Facade as IlluminateFacade;
+use Illuminate\Support\Facades\Facade;
 
-class Facade extends IlluminateFacade
+class Zillow extends Facade
 {
     /**
      * Get the registered name of the component.
